add trait for OpenAmDAO
